pyproject.toml
[project]
name = "minesweepervariants"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13,<3.15"
dependencies = [
    "cython (>=3.1.1,<3.2.0)",
    "ortools (>=9.13.4784,<9.14.0)",
    "numpy (>=2.3.0,<2.4.0)",
    "pandas (>=2.3.0,<2.4.0)",
    "pillow (>=11.3.0,<11.4.0)",
    "pyyaml (>=6.0.2,<6.1.0)",
    "flask (>=3.1.1,<3.2.0)",
    "flask-cors (>=6.0.1,<7.0.0)",
    "cryptography (>=45.0.6,<46.0.0)",
    "waitress (>=3.0.2,<3.1.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pyinstaller = "^6.15.0"

